+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## [1.0.0] (IN PROGRESS)
 
 * [UIOR-1080](https://issues.folio.org/browse/UIOR-1080) Setup new UI project `ui-consortia-settings`.
+* [UIOR-1081](https://issues.folio.org/browse/UIOR-1081) Add `Membership` section to the `Consortium` settings page.
+* [UIOR-1082](https://issues.folio.org/browse/UIOR-1082) Add controlled vocabulary component to render consortium members.